split memory cache into separate files
Carves out the MemoryCache struct and cachesForNode() function into
separate files.

Issue #33
